chapter_011/requirements.txt
# LangChain ecosystem
langchain==1.2.0
langchain-core==1.2.15
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-openai==1.1.10
langchain-google-genai==4.1.2
langchain-anthropic==1.3.4
langchain-classic==1.0.0

# LLM API clients
openai>=2.8.1
anthropic==0.83.0

# Tokenizer
tiktoken==0.12.0

# Streamlit (latest installed)
streamlit==1.51.0
streamlit-feedback==0.1.4

# Environment
python-dotenv==1.2.1

# Web / Crawler / Parsing
requests==2.32.5
beautifulsoup4==4.14.2
ddgs==9.10.0
html2text==2025.4.15
lxml==6.0.2
lxml_html_clean==0.4.3
readability-lxml==0.8.4.1

# Youtube
youtube-transcript-api==1.2.3
pytube==15.0.0

# PDF / Vector DB
PyMuPDF==1.26.6
faiss-cpu==1.13.0

# LangSmith
langsmith>=0.6.0

# File & BigQuery
python-magic==0.4.27
google-cloud-bigquery==3.38.0
db-dtypes==1.4.4

